app/requirements.txt
Django>=5.2
python-dotenv
psycopg2-binary>=2.9